reworks LogIn and SignUp,updated models/app/seed/alembic

+++ server/app.py
@@ -1,23 +1,30 @@
 #!/usr/bin/env python3
 
 # Standard library imports
-from flask import Flask
+import os
+BASE_DIR = os.path.abspath(os.path.dirname(__file__))
+DATABASE = os.environ.get("DB_URI", f"sqlite:///{os.path.join(BASE_DIR, 'instance/app.db')}")
+
+from flask import Flask, jsonify, make_response, request
+
 from flask_migrate import Migrate
-from flask_restful import Api
+from flask_restful import Api, Resource
 from flask_sqlalchemy import SQLAlchemy
-from flask_restful import Resource
-from flask import request, jsonify, make_response, session
+
+from models import db, User, Job, Company
 
 # Instantiate Flask app
 app = Flask(__name__)
 
 # Configure app
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///workwander.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = DATABASE
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.json.compact = False
 
-# Instantiate db
-db = SQLAlchemy(app)
+
 migrate = Migrate(app, db)
+
+db.init_app(app)
 
 # Instantiate REST API
 api = Api(app)
@@ -25,7 +32,6 @@ api = Api(app)
 # -------bcrypt----------------
 
 # Import models after initializing db
-from models import User, Job, AppliedJob, Salary, CompanyReview, Company
 
 # Views go here!
 
@@ -35,74 +41,54 @@ def index():
     return '<h1>WorkWander</h1>'
 
 
-class Users(Resource):
+class UserList(Resource):
     def get(self):
-        users = [user.tp_dict() for user in User.query.all()]
-        return users
+        users = [user.to_dict() for user in User.query.all()]
+        return make_response(jsonify(users), 200)
+    
+api.add_resource(UserList, '/users')
+    
+class UserListById(Resource):
+    def get(self, id):
+        user = User.query.get(id)
+        if not user:
+            return make_response({"error": "404: Camper not found."}, 404)
+        return make_response(jsonify(user.to_dict()), 200)
+    
+api.add_resource(UserListById, '/users/<int:id>')
 
 
-api.add_resource(Users, '/users')
 
-
-class Jobs(Resource):
+class JobList(Resource):
     def get(self):
-        jobs = [job.tp_dict() for job in Job.query.all()]
-        return jobs
+        job_list = [job.to_dict() for job in Job.query.all()]
+        return make_response(jsonify(job_list), 200)
     
     def delete(self):
-        job = Job.query.get(request.json['id'])
+        job_id = request.json.get('id')
+        if job_id is None:
+            return make_response(jsonify({'error': "400: Missing 'id' parameter"}), 400)
+
+        job = Job.query.get(job_id)
+        if job is None:
+            return make_response(jsonify({'error': "404: Job not found"}), 404)
+
         db.session.delete(job)
         db.session.commit()
-        return job.to_dict()
+        return {}, 200
 
 
-api.add_resource(Jobs, '/jobs')
+api.add_resource(JobList, '/jobs')
 
 
-class AppliedJobs(Resource):
+
+class CompanyList(Resource):
     def get(self):
-        applied_jobs = [applied_job.tp_dict() for applied_job in AppliedJob.query.all()]
-        return applied_jobs
+        companies = [company.to_dict() for company in Company.query.all()]
+        return make_response(jsonify(companies), 200)
 
-
-api.add_resource(AppliedJobs, '/applied-jobs')
-
-
-class Salaries(Resource):
-    def get(self):
-        salaries = [salary.tp_dict() for salary in Salary.query.all()]
-        return salaries
-
-
-api.add_resource(Salaries, '/salaries')
-
-
-class CompanyReviews(Resource):
-    def get(self):
-        company_reviews = [company_review.tp_dict() for company_review in CompanyReview.query.all()]
-        return company_reviews
-    
-    def post(self):
-        company_review = CompanyReview(
-            user_id=request.json['user_id'],
-            company_id=request.json['company_id'],
-            review=request.json['review'],
-            rating=request.json['rating']
-        )
-
-
-api.add_resource(CompanyReviews, '/company-reviews')
-
-
-class Companies(Resource):
-    def get(self):
-        companies = [company.tp_dict() for company in Company.query.all()]
-        return companies
-
-
-api.add_resource(Companies, '/companies')
+api.add_resource(CompanyList, '/companies')
 
 
 if __name__ == '__main__':
-    app.run(debug=True)
-
+    app.run(port=5000, debug=True)
